Be more flexible to accept new commands
One could argue that the initializer is now simpler, alas there is now
metaprogramming. Or as I call it: there, I used ruby.

Also, whoever says 'ActiveSupport/String-Additions' will be confronted
with a massive memory footprint and might be tied to railroad-tracks.
Yes, I realize that railties are not part of AS...